LibSyntax: Move Language enum into its own files
