+++ Userland/Libraries/LibSyntax/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(SOURCES
     Highlighter.cpp
+    Language.cpp
 )
 
 serenity_lib(LibSyntax syntax)
